Move Metadata interfaces to Application layer
The interfaces, contracting the behaviour of the fetcher, parser and
certificate parser have been moved to the Application layer.
